requirements-dev.txt
-r requirements.txt
debugpy==1.8.0
pylint
pylint-django
